Move back to routing package (for now)
