+++ core/src/main/java/com/graphhopper/routing/QueryGraphTurnExt.java
@@ -16,7 +16,7 @@
  *  limitations under the License.
  */
 
-package com.graphhopper.routing.query_graph;
+package com.graphhopper.routing;
 
 import com.carrotsearch.hppc.IntArrayList;
 import com.graphhopper.storage.Graph;
