+++ core/src/main/java/com/graphhopper/routing/template/RoutingTemplate.java
@@ -19,8 +19,8 @@ package com.graphhopper.routing.template;
 
 import com.graphhopper.routing.AlgorithmOptions;
 import com.graphhopper.routing.Path;
+import com.graphhopper.routing.QueryGraph;
 import com.graphhopper.routing.RoutingAlgorithmFactory;
-import com.graphhopper.routing.query_graph.QueryGraph;
 import com.graphhopper.routing.util.FlagEncoder;
 import com.graphhopper.storage.index.QueryResult;
 import com.graphhopper.util.PathMerger;
