+++ core/src/main/java/com/graphhopper/routing/util/TestAlgoCollector.java
@@ -19,7 +19,6 @@ package com.graphhopper.routing.util;
 
 import com.graphhopper.PathWrapper;
 import com.graphhopper.routing.*;
-import com.graphhopper.routing.query_graph.QueryGraph;
 import com.graphhopper.routing.weighting.TurnWeighting;
 import com.graphhopper.storage.CHGraph;
 import com.graphhopper.storage.Graph;
