+++ core/src/test/java/com/graphhopper/routing/template/RoundTripRoutingTemplateTest.java
@@ -20,11 +20,7 @@ package com.graphhopper.routing.template;
 import com.carrotsearch.hppc.IntArrayList;
 import com.graphhopper.GHRequest;
 import com.graphhopper.GHResponse;
-import com.graphhopper.routing.AlgorithmOptions;
-import com.graphhopper.routing.AlternativeRouteTest;
-import com.graphhopper.routing.Path;
-import com.graphhopper.routing.RoutingAlgorithmFactorySimple;
-import com.graphhopper.routing.query_graph.QueryGraph;
+import com.graphhopper.routing.*;
 import com.graphhopper.routing.util.*;
 import com.graphhopper.routing.weighting.FastestWeighting;
 import com.graphhopper.routing.weighting.Weighting;
